spliting sql_engine tests
